Fix compiler warnings and clean up unused variables across modules
- Cast variables to fix type mismatches and suppress warnings in dict-dsl.c and dict-mdx.c
- Silence unused parameters and remove unused variables in dict-mdx.c and search-utils.c
- Replace deprecated g_qsort_with_data with g_sort_array in settings-dialog.c

+++ src/dict-dsl.c
@@ -140,7 +140,7 @@ static void dsl_parse_header(const char *path, char **out_name, char **out_sourc
     }
 
     unsigned char bom[4];
-    int bom_len = is_dz ? gzread(gz, bom, 4) : fread(bom, 1, 4, f);
+    size_t bom_len = is_dz ? (size_t)gzread(gz, bom, 4) : fread(bom, 1, 4, f);
     int skip = 0, bom_is_utf16 = 0, utf16_be = 0;
     if (bom_len >= 3 && bom[0] == 0xEF && bom[1] == 0xBB && bom[2] == 0xBF) {
         skip = 3;
